lib/mysql_driver/requirements.txt
#mysql-connector-python
mysql-connector